Create skills JSON lists and break home.tsx out into multiple components
